pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cloudflare_access_jwt"]

[project]
name = "fastapi-cloudflare-access-jwt"
version = "0.1.0"
description = "FastAPI middleware for Cloudflare Access JWT authentication"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
authors = [{ name = "Alexandru Rosianu", email = "me@aluxian.com" }]
keywords = ["fastapi", "cloudflare", "jwt", "authentication", "middleware"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: FastAPI",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP :: WSGI :: Middleware",
    "Topic :: Security",
]
dependencies = [
    "fastapi[standard]>=0.110.0",
    "PyJWT>=2.8.0",
    "cryptography>=42.0.5",
    "loguru>=0.7.2",
    "requests>=2.31.0",
]

[project.urls]
Homepage = "https://github.com/aluxian/fastapi-cloudflare-access-jwt"
Repository = "https://github.com/aluxian/fastapi-cloudflare-access-jwt.git"
Issues = "https://github.com/aluxian/fastapi-cloudflare-access-jwt/issues"
Changelog = "https://github.com/aluxian/fastapi-cloudflare-access-jwt/releases"

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.3",
    "requests-mock>=1.11.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.7.2",
]
